* master:
  service failure delete route
  testing
  test
  deleting failures API fix - postman API update
  additional testing
  testing - cleanup

+++ handlers/services.go
@@ -326,8 +326,12 @@ func servicesDeleteFailuresHandler(w http.ResponseWriter, r *http.Request) {
 	}
 	vars := mux.Vars(r)
 	service := core.SelectService(utils.ToInt(vars["id"]))
+	if service == nil {
+		sendErrorJson(errors.New("service not found"), w, r)
+		return
+	}
 	service.DeleteFailures()
-	ExecuteResponse(w, r, "services.gohtml", core.CoreApp.Services, "/services")
+	sendJsonAction(service, "delete_failures", w, r)
 }
 
 func apiServiceFailuresHandler(w http.ResponseWriter, r *http.Request) {
